refactor(viewer): rename compile script to compile-shader
